인증블록 수정
1. 인증블록 다른 봇 불러오기 허용
2. 봇 수정 시 기존 인증블럭 결과 삭제

+++ api-server/src/main/java/kr/co/eicn/ippbx/server/repository/eicn/WebchatBotDisplayRepository.java
@@ -40,8 +40,9 @@ public class WebchatBotDisplayRepository extends EicnBaseRepository<WebchatBotDi
     }
 
     public void deleteByBlockIdList(List<Integer> blockIdList) {
-        dsl.deleteFrom(WEBCHAT_BOT_DISPLAY)
-                .where(WEBCHAT_BOT_DISPLAY.BLOCK_ID.in(blockIdList))
-                .execute();
+        if (blockIdList != null && blockIdList.size() > 0)
+            dsl.deleteFrom(WEBCHAT_BOT_DISPLAY)
+                    .where(WEBCHAT_BOT_DISPLAY.BLOCK_ID.in(blockIdList))
+                    .execute();
     }
 }
